Bug fix for 1.0 compatible expressions (#144)
Fixing bug that was introduced for handling 1.0 compatible expressions. This resolves the issue reported in https://github.com/aws-cloudformation/cloudformation-guard/issues/142 for rulegen clauses as well.

Issue:
With 1.0 expressions of the form Statement.*.Principal == '*' was allowed. To be backwards compatible with these type of rules support was added to 2.0. However, there are cases that can lead to query results having mixed values of scalars and list of scalars returned. Here is an example of such a case.

Statement:
  - Principal: '*'
  - Principal: ['*', 'aws:arn']
The query results in

- *
- ['*', 'aws:arn']

This often happens with IAM policies as attributes like Action, Resource, and Principal can be provided both as an array form or single value form. If the Guard 2.0 rule was written as Statement[*].Principal[*], that would result in the right evaluation. However, Guard 2.0 did add support for 1.0 compatible syntax Statement.*.Principal == '*'. 

Fix:
When queries result in mixed valued results we flatten and check the final results. When the queries are not these mismatched type check, then we delegate to the default check, which addressed 142